+++ src/pages/app.js
@@ -1,8 +1,8 @@
 import React from "react"
 import { Router } from "@reach/router"
 import Layout from "../components/layout"
-import PrivateRoute from "../components/privateRoute"
-import Login from "../components/login"
+import PrivateRoute from "../components/authentication/privateRoute"
+import Login from "../components/authentication/login"
 import postJob from "../components/postJob";
 import businessProfile from "../components/business_profile/businessProfile"
 import Main from '../components/Main'
